add validation logic for admin checking & unit test case

+++ src/main/kotlin/logic/usecases/project/DeleteProjectUsecase.kt
@@ -1,27 +1,22 @@
 package logic.usecases.project
 
+import logic.entities.User
 import logic.exception.PlanMateException
+import logic.exception.PlanMateException.AuthorizationException.AdminPrivilegesRequiredException
 import logic.exception.PlanMateException.DataSourceException.EmptyDataException
 import logic.exception.PlanMateException.ValidationException.InvalidProjectIDException
 import logic.repository.ProjectRepository
+import logic.usecases.ValidateInputUseCase
 import java.util.UUID
 
-class DeleteProjectUseCase(val projectRepository: ProjectRepository){
+class DeleteProjectUseCase(val projectRepository: ProjectRepository, private val user: User, private val validateInputUseCase: ValidateInputUseCase ){
 
     fun deleteProject(projectId: String): Boolean {
-        validateId(projectId)
+        if (projectId.isEmpty() || projectId.isBlank()) throw EmptyDataException
+        if(! validateInputUseCase.isValidUUID(projectId))throw InvalidProjectIDException
+        if(!user.isAdmin) throw AdminPrivilegesRequiredException
         return projectRepository.deleteProject(UUID.fromString(projectId))
     }
 
-    private fun validateId(id: String) {
-        if(id.isEmpty() || id.isBlank()){
-            throw EmptyDataException
-        }
-        try {
-            UUID.fromString(id)
-        }catch (e: IllegalArgumentException){
-            throw InvalidProjectIDException
-        }
-    }
 
 }
